Consolidate graph embedding and umap
