--- graphrag/index/operations/layout_graph/methods/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 Microsoft Corporation.
-# Licensed under the MIT License
-
-"""Graph Layout Methods."""
